🚚 consolidate iucn helpers, clean up client exports
