+++ packages/geoprocessing/client-core.ts
@@ -4,17 +4,13 @@
 export * from "./src/types";
 
 // Helpers - not all of them
-export * from "./src/helpers/units";
-export * from "./src/helpers/keyBy";
 export * from "./src/helpers/groupBy";
-export * from "./src/helpers/string";
+export * from "./src/helpers/keyBy";
 export * from "./src/helpers/number";
+export * from "./src/helpers/string";
+export * from "./src/helpers/sketch";
 export * from "./src/helpers/types";
+export * from "./src/helpers/units";
+
 export * from "./src/metrics/helpers";
-export {
-  toSketchArray,
-  toNullSketch,
-  toNullSketchArray,
-  getUserAttribute,
-  getJsonUserAttribute,
-} from "./src/helpers/sketch";
+export * from "./src/iucn/helpers";
